Create ALONE NUMBER
Problem: having N = 2n + 1 intergers, in which 1 number not have pair (such as 1 & 1, 2 & 2, ...). The mission is to find that number in the array

Input : 
5
1 1 2 2 3

Output: 3


HOW TO SOLVE:
- using XOR (eXclusive OR) (^ in calling in c++)
- XOR: return 1 if 1 of the 2 respective bits is 1 and 0 if both are 1 / 0

Step 1: set a RESULT variable and set  RESULT = 0
Step 2: Read and XOR each part  with result. If XOR of the 2 numbers is identical => result = 0
Step 3: At last, the 'lonely  number' is the result